+++ lib/src/sound/events/sound_position.dart
@@ -13,23 +13,28 @@ const unpanned = SoundPosition();
 
 /// A sound which should be panned.
 ///
-/// You can set the left right panning with the [azimuthOrScalar] property.
-class SoundPositionPanned extends SoundPosition {
+/// You can set the left right panning with the [scalar] property.
+class SoundPositionScalar extends SoundPosition {
   /// Create a panned position.
-  const SoundPositionPanned({this.azimuthOrScalar = 0.0, this.elevation});
+  const SoundPositionScalar({this.scalar = 0.0});
 
   /// The left right balance of this sound.
-  ///
-  /// If [elevation] is `null`, then this value is treated as a scalar.
-  ///
-  /// Otherwise, this value is treated as an azimuth value.
-  final double azimuthOrScalar;
+  final double scalar;
+}
 
-  /// The elevation to use.
-  ///
-  /// If this value is not `null`, then [azimuthOrScalar] is treated as an
-  /// azimuth value.
-  final double? elevation;
+/// A sound which should be panned.
+///
+/// You can set the left right panning with the [azimuth] and [elevation]
+/// properties.
+class SoundPositionAngular extends SoundPosition {
+  /// Create a panned position.
+  const SoundPositionAngular({this.azimuth = 0.0, this.elevation = 0.0});
+
+  /// The azimuth of this sound.
+  final double azimuth;
+
+  /// The elevation of this sound.
+  final double elevation;
 }
 
 /// A sound which should be positioned in 3d space.